Align project structure with other server implementations
- Organize into separate server and deployer modules
  and align structures with other implementations.
- Fixes #76
