+++ spring-cloud-dataflow-server-yarn/src/main/java/org/springframework/cloud/dataflow/server/yarn/YarnDataFlowServer.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2015-2016 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.cloud.dataflow.server.yarn;
+
+import org.springframework.cloud.dataflow.admin.AdminApplication;
+import org.springframework.cloud.dataflow.admin.EnableDataFlowServer;
+
+/**
+ * Bootstrap class for Spring Cloud Data Flow Admin on Apache Hadoop YARN.
+ *
+ * @author Mark Fisher
+ * @author Janne Valkealahti
+ */
+@EnableDataFlowServer
+public class YarnDataFlowServer {
+
+	public static void main(String[] args) {
+		AdminApplication.main(args);
+	}
+}
