+++ spring-cloud-dataflow-server-yarn/src/main/resources/admin.yml
