+++ spring-cloud-dataflow-server-yarn/src/main/resources/stream.yml
