+++ spring-cloud-dataflow-server-yarn/src/main/resources/task.yml
